test(wtr): remove LWC import from utils
It breaks hydration tests if there's an import from LWC.
