+++ packages/@lwc/integration-not-karma/test/mixed-shadow-mode/reporting/index.spec.js
@@ -6,7 +6,7 @@ import Light from 'x/light';
 import {
     attachReportingControlDispatcher,
     detachReportingControlDispatcher,
-} from '../../../helpers/utils.js';
+} from '../../../helpers/reporting-control.js';
 
 // Should be kept in sync with the enum in vm.ts
 const ShadowMode = {
