+++ packages/@lwc/integration-not-karma/test/mixed-shadow-mode/shadowSupportModeReporting/index.spec.js
@@ -7,7 +7,7 @@ import NativeOnly from 'x/native';
 import {
     attachReportingControlDispatcher,
     detachReportingControlDispatcher,
-} from '../../../helpers/utils.js';
+} from '../../../helpers/reporting-control.js';
 
 /**
  * These tests must be the first ones to generate the component def for the components they use.
